Tic Tac Toe windows form app
here's a windows form in the meantime. It has two options so far one for 2 players and one easy level to play against the computer using random. I didn't use Ari's code yet for the checkWin method if you could do it please do it just has to converted to C#.